added cli docs, changed to list-required-secrets
